Quick look into posts.db
Using knowledge of database for login, try making a database using posts.db for content. The idea is to display the content on the homepage (once logged in) with a thumbnail in the black container and title (maybe description) in the red container surrounding it (but to the right). For each entry in the database this will be done for. Input boxes on new guide need to be adjusted and method for actually getting the data out of the db has not been started